refactor: 불필요한 렌더링 요소 제거 및 코드 가독성 개선
